Add option to draw ecal halo in all 3D view types.
