+++ Fireworks/Core/interface/FWISpyView.h
@@ -24,7 +24,6 @@
 #include "Fireworks/Core/interface/FW3DViewBase.h"
 
 // forward declarations
-class TEveBoxSet;
 
 class FWISpyView : public FW3DViewBase
 {
@@ -34,7 +33,6 @@ public:
    virtual void setContext(const fireworks::Context& x);
 
    virtual void populateController(ViewerParameterGUI&) const;
-   void showEcalBarrel(bool);
    // ---------- const member functions ---------------------
 
    // ---------- static member functions --------------------
@@ -46,10 +44,7 @@ private:
 
    const FWISpyView& operator=(const FWISpyView&); // stop default
 
-   // ---------- member data --------------------------------
-   TEveBoxSet* m_ecalBarrel;
-   FWBoolParameter m_showEcalBarrel;
+   // ---------- member data --------------------------------};
 };
-
 
 #endif
